Allow passing in current lat/lon of plane
Simply add as path components to the API call.
Both versions are supported for now.

Add a little C++ helper app to do some spherical geometry for us.
This allows the code to guess wether the current position of the
plane is plausible, given the declared callsign / route (well,
really it's the other way around... given the position, is the
route plausible).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>