Add routeset API
This endpoint expects a POST request with JSON data that includes an
array of structures consisting of callsign, lat, lon. It then checks
for route information for each of those callsigns (and adds a
plausibility check based on the given position).
The result is returned as a JSON encoded array of routes.

This setup should create significantly less API load if paired with
a frontend implementation that only checks every 5-10 seconds.

A pull request for tar1090 that does just that is here:
https://github.com/wiedehopf/tar1090/pull/234

All this is only "lightly tested". Also, I'm not a Python programmer,
so this might benefit from some careful review. And I'm very open to
edits / suggestions.

Also, note that this is back to using the CORS setup (as with the
changes to tar1090 this appears to be needed again) - and right now this
is globally implemented, which you obviously don't want.

I'm running out of brain cells to remember how I initially implemented
this to only use it for this API call. So please add that before
merging.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>