fix CSV parsing
The routes CSV contains airport names with commas, correctly quoted
with double quotes. Using split(',') on those lines will throw an error.
But the CSV parser can handle a single line if you ask nicely enough.

Hat tip to Katia for explaining most of this to me and to
https://stackoverflow.com/questions/35822822/parse-a-single-csv-string
for the right approach to use the csv.reader().

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>